Reworking database purging in fixtures
Associated entities are no longer deleted unless they are declared as
part of the fixtures, or all fixtures are to run (how doctrine-bundle
behaves).

This commit replaces the hacky mapping driver with more purger logic.
Gedmo extensions are very particular about the classnames of the mapping
drivers, and this driver was incompatible.